+++ src/auth/basic.js
@@ -19,10 +19,11 @@ export default class BasicAuth extends Auth {
   }
 
   getButtonTitle() {
-  return i18n.global.t('authentication.button.title');
+    return i18n.global.t('authentication.button.title');
   }
 
-  async logout(/*credentials*/) {
+  async logout(credentials) {
+    await super.logout(credentials);
     if (this.router.currentRoute.value.name !== 'logout') {
       this.router.push('/auth/logout');
     }